hid api key moved test folded,refactored App.js main.js
