+++ src/App.js
@@ -3,7 +3,8 @@ import './main-css/App.css'
 import Header from './Header'
 import Main from './Main'
 import $ from 'jquery'
-
+var key = require("./apiKey")
+var myKey = key.key
 
 
 class App extends Component {
@@ -24,7 +25,8 @@ class App extends Component {
       condition: '',
       high: '',
       low: '',
-      summary: ''
+      summary: '',
+      apiSource:"http://api.wunderground.com/api"
     }
   }
 
@@ -37,7 +39,7 @@ class App extends Component {
         navigator.geolocation.getCurrentPosition(this.findLocation.bind(this))
       } else {
         $.getJSON(
-          `http://api.wunderground.com/api/3d896652346518f2/forecast/hourly/hourly10day/q/${State}/${city}.json`
+          `${this.state.apiSource}/${myKey}/forecast/hourly/hourly10day/q/${State}/${city}.json`
         ).then(weather => this.apiEdit(weather))
       }
     })
@@ -47,12 +49,12 @@ class App extends Component {
     let latitude = position.coords.latitude
     let longitude = position.coords.longitude
     $.getJSON(
-      `http://api.wunderground.com/api/3d896652346518f2/geolookup/q/${latitude},${longitude}.json`
+      `${this.state.apiSource}/${myKey}/geolookup/q/${latitude},${longitude}.json`
 
     ).then(weather => {
       let zip = weather.location.zip
       $.getJSON(
-        `http://api.wunderground.com/api/3d896652346518f2/forecast/hourly/hourly10day/conditions/q/${zip}.json`
+        `${this.state.apiSource}/${myKey}/forecast/hourly/hourly10day/conditions/q/${zip}.json`
       ).then(locationWeather => {
         this.setState({city: locationWeather.current_observation.display_location.city, State: locationWeather.current_observation.display_location.state_name})
         localStorage.setItem('city', locationWeather.current_observation.display_location.city)
@@ -130,7 +132,7 @@ class App extends Component {
           localStorage.setItem('city', probableLocation[0])
           localStorage.setItem('State', probableLocation[1])
           $.get(
-            `http://api.wunderground.com/api/3d896652346518f2/forecast/hourly/hourly10day/conditions/q/${probableLocation[1]}/${probableLocation[0]}.json`
+            `${this.state.apiSource}/${myKey}/forecast/hourly/hourly10day/conditions/q/${probableLocation[1]}/${probableLocation[0]}.json`
           ).then(weather => {
             this.apiEdit(weather)
             this.setState({currentState: probableLocation[1]})
@@ -142,7 +144,7 @@ class App extends Component {
       localStorage.setItem('State', this.state.State)
 
       $.get(
-        `http://api.wunderground.com/api/3d896652346518f2/forecast/hourly/hourly10day/conditions/q/${this.state.State}/${this.state.city}.json`
+        `${this.state.apiSource}/${myKey}/forecast/hourly/hourly10day/conditions/q/${this.state.State}/${this.state.city}.json`
       ).then(weather => this.apiEdit(weather)).catch(() => {
         alert('Sorry Something Went Wrong ☹️, please enter a city, zipcode, or state')
       })
